[+] Ability to choose torrent file's character set encoding in torrent open dialog.
git-svn-id: https://flush.svn.sourceforge.net/svnroot/flush/trunk@50 21f5d154-ea94-4097-a85b-94d12d4401c9

+++ src/mlib/misc.cpp
@@ -464,7 +464,7 @@ void close_all_fds(void) throw(m::Exception)
 	if(getrlimit(RLIMIT_OFILE, &limits))
 		M_THROW(__("Can't get max opened files limit: %1.", EE(errno)));
 
-	for(int fd = STDERR_FILENO + 1; fd < limits.rlim_max; fd++)
+	for(int fd = STDERR_FILENO + 1; fd < (int) limits.rlim_max; fd++)
 		close(fd);
 }
